src/server_py/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
httpx>=0.25.0
psycopg2-binary>=2.9.9
alembic>=1.12.1
pytest>=7.4.3
asyncpg>=0.28.0
email-validator>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.6
uuid>=1.30
fastapi-pagination>=0.12.10
# RAG dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
# LangChain dependencies
langchain>=0.0.267
langchain-community>=0.0.10
langchain-ollama>=0.3.0
langchainhub>=0.1.13
huggingface-hub>=0.17.3
transformers>=4.33.0